Windows mingw makefile
To use this makefile erase MINGW from the name and make sure it is the only makefile in the directory containing the game. After that go to your directory in your terminal thing and type "make" and assuming everything works properly it should comipile as "Game.exe". I am using MSYS as a unix based terminal that you can get with to complement the mingw compiler and stuff, but really I think you just need a terminal with "make".